fix: убирает все функции иp main.js
